Fetch results eagerly from coordinator
Before, the query results would only begin to be fetched once the client was
redirected to the executing query endpoint.  This introduces latency and may
not be acceptable for very low latency use cases.
